Update sensor.py (#34)
I am a big fan of this project, and using this to compare my installation without a battery and with this virtual battery. To see in a year if I want to buy a home battery.
I am currently unable to compare due to the rounding to 2. By rounding set to 3, which is a standard for kw, I am able to convert back to watts and compare with other meters containing the 3 digits precision. In Battery Sim 1.999 kW = 2kW, meaning 2000W. But I want to read 1999W instead. 

Pleas accept this modification. And a gain a big thanks for this project.